requirement.txt
streamlit
pandas
numpy
plotly
scikit-learn
openai
python-dotenv

